[Feature] 쿼리 updateOne, create, deleteOne 구현 (#15)
* refactor: :hammer: module pattern 적용하여 캡슐화 #11

* bug: :bug: test 에러 해결
버튼 누를 때마다 2^으로 이벤트가 붙음 , 이벤트 등록 위치 수정

* feat: :sparkles: database 연결 #7

* feat: :sparkles: database 연결 test 함수 #7

* feat: :sparkles: updateOne, upateAll, deleteOne 구현 #7
query 함수 구현 및 테스트 파일 추가

* feat: :sparkles: updateOne 성공시 반영전 row데이터 반환

* feat: :sparkles: deleteOne, create  성공시 의미있는 반환 값 return 구현
delete는 이전 값 create는  생성값

+++ backend/src/lib/query.js
@@ -8,36 +8,72 @@ const dbtest = async () => {
 
 // prettier-ignore
 const find = (table, condition, order) => {
-  //join 고려해서 맡아서 짜오기 시간 조건도 고려해야합니다. ;
+ 
+};
+
+const findOne = async (table, condition) => {
+  try {
+    const conditionTemplate = createEqualsTemplate(condition);
+    const seleteQuery = `SELECT * FROM ${table} WHERE ${conditionTemplate}`;
+    const [rows] = await getDB().query(seleteQuery);
+    return rows.length === 0 ? null : rows[0];
+  } catch (err) {
+    console.err(err);
+    return null;
+  }
 };
 
 const updateOne = async (table, condition, updateMap) => {
-  const conditionTemplate = createEqualsTemplate(condition);
-  const updateTemplate = createEqualsTemplate(updateMap);
-  const updateQuery = ` UPDATE ${table} SET ${updateTemplate} WHERE ${conditionTemplate}`;
-  return await getDB().execute(updateQuery);
+  try {
+    const beforeUpdateData = await findOne(table, condition);
+    if (!beforeUpdateData) return null;
+    const conditionTemplate = createEqualsTemplate(condition);
+    const updateTemplate = createEqualsTemplate(updateMap);
+    const updateQuery = ` UPDATE ${table} SET ${updateTemplate} WHERE ${conditionTemplate}`;
+    const [row] = await getDB().query(updateQuery);
+    return row.affectedRows === 1 ? beforeUpdateData : null;
+  } catch (err) {
+    console.err(err);
+    return null;
+  }
 };
 
 const updateAll = (table, condition, updateMap) => {};
 
 const deleteOne = async (table, condition) => {
-  const conditionTemplate = createEqualsTemplate(condition);
-  const deleteQuery = `DELETE FROM ${table} WHERE ${conditionTemplate};`;
-  return await getDB().execute(deleteQuery);
+  try {
+    const beforeDeleteData = await findOne(table, condition);
+    if (!beforeDeleteData) return null;
+    const conditionTemplate = createEqualsTemplate(condition);
+    const deleteQuery = `DELETE FROM ${table} WHERE ${conditionTemplate};`;
+    const [row] = await getDB().query(deleteQuery);
+    return row.affectedRows === 1 ? beforeDeleteData : null;
+  } catch (err) {
+    console.err(err);
+    return null;
+  }
 };
 
 const deleteAll = (table, condition) => {};
 
 const create = async (table, createMap) => {
-  const keys = [],
-    values = [];
-  Object.entries(createMap).forEach((item) => {
-    const [key, value] = item;
-    keys.push(key);
-    values.push(convertType(value));
-  });
-  const createQuery = `INSERT INTO ${table} ( ${keys.join(',')} ) values ( ${values.join(',')} )`;
-  return await getDB().execute(createQuery);
+  try {
+    const keys = [],
+      values = [];
+    Object.entries(createMap).forEach((item) => {
+      const [key, value] = item;
+      keys.push(key);
+      values.push(convertType(value));
+    });
+    const createQuery = `INSERT INTO ${table} ( ${keys.join(',')} ) values ( ${values.join(',')} )`;
+    const [row] = await getDB().execute(createQuery);
+    if (row.affectedRows !== 1) return null;
+    const createData = await findOne(table, { id: row.insertId });
+    return createData;
+  } catch (err) {
+    console.err(err);
+    return null;
+  }
 };
 
 const createEqualsTemplate = (condition) => {
@@ -58,8 +94,6 @@ module.exports = {
   dbtest,
   find,
   updateOne,
-  updateAll,
   deleteOne,
-  deleteAll,
   create,
 };
